+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - added protocol list to data source model in myems-admin
-- 
+- added microgrid dashboard and energy storage power station dashboard into myems-web
 ### Changed
 - updated space export, import and clone functions in myems-api
 - updated microgrid reporting in myems-api and myems-web